rid of costs and visited in Vertex
